requirements.txt
opencv-python
numpy
pickle-mixin
torch
torchvision
scipy
keras-facenet
mediapipe
ultralytics
deep_sort_realtime
mysql-connector-python
csv
argparse
queue
time
datetime
logging
watchdog